better collission detection with straal 2 more wall types food spreader test
git-svn-id: svn://84.194.112.164/bengine@229 2ee6282c-0497-4b59-8b7c-bcac2105c061

+++ src/scenes/entities/worldb.h
@@ -40,8 +40,8 @@ class WorldB
 
 		unsigned long		currentCritterID;
 
-		unsigned int		sizeX;
-		unsigned int		sizeY;
+		float			sizeX;
+		float			sizeY;
 		float			freeEnergy;
 		//float			freeEnergyInfo;
